# docs/swagger/sql-engine.openapi.json
{"openapi":"3.1.0","info":{"title":"SQL Engine API","description":"OLAP 查询层 — 模板 SQL / 自然语言 / 用户分组，与 Doris 解耦","version":"1.1.0","x-services":{"id-mapping":"http://localhost:8001/openapi.json","sql-engine":"http://localhost:8002/openapi.json"}},"paths":{"/health":{"get":{"tags":["系统"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/templates":{"get":{"tags":["模板查询"],"summary":"List Templates","operationId":"list_templates_templates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/nl-query":{"post":{"tags":["自然语言"],"summary":"Natural Language Query","description":"自然语义查询：DeepSeek/规则 → 模板 + 参数 → 执行（支持用户、分组）","operationId":"natural_language_query_nl_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NlQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NlQueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query/{template_name}":{"post":{"tags":["模板查询"],"summary":"Run Query","operationId":"run_query_query__template_name__post","parameters":[{"name":"template_name","in":"path","required":true,"schema":{"type":"string","title":"Template Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["模板查询"],"summary":"Run Query Get","operationId":"run_query_get_query__template_name__get","parameters":[{"name":"template_name","in":"path","required":true,"schema":{"type":"string","title":"Template Name"}},{"name":"tenant_id","in":"query","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"one_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"One Id"}},{"name":"phone","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"}},{"name":"form_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Form Id"}},{"name":"channel_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Channel Type"}},{"name":"channel_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Channel Id"}},{"name":"group_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Group Id"}},{"name":"group_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Group Code"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups":{"post":{"tags":["用户分组"],"summary":"Create Group","operationId":"create_group_groups_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{tenant_id}":{"get":{"tags":["用户分组"],"summary":"List Groups","operationId":"list_groups_groups__tenant_id__get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupResponse"},"title":"Response List Groups Groups  Tenant Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{tenant_id}/{group_id}":{"get":{"tags":["用户分组"],"summary":"Get Group","operationId":"get_group_groups__tenant_id___group_id__get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{tenant_id}/code/{group_code}":{"get":{"tags":["用户分组"],"summary":"Get Group By Code","operationId":"get_group_by_code_groups__tenant_id__code__group_code__get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"group_code","in":"path","required":true,"schema":{"type":"string","title":"Group Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{tenant_id}/{group_id}/members":{"post":{"tags":["用户分组"],"summary":"Add Member","operationId":"add_member_groups__tenant_id___group_id__members_post","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberAdd"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["用户分组"],"summary":"List Members","operationId":"list_members_groups__tenant_id___group_id__members_get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{tenant_id}/{group_id}/members/{one_id}":{"delete":{"tags":["用户分组"],"summary":"Remove Member","operationId":"remove_member_groups__tenant_id___group_id__members__one_id__delete","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"one_id","in":"path","required":true,"schema":{"type":"integer","title":"One Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{tenant_id}/{one_id}":{"get":{"tags":["用户"],"summary":"Get User","description":"用户画像 + 所属分组","operationId":"get_user_users__tenant_id___one_id__get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"one_id","in":"path","required":true,"schema":{"type":"integer","title":"One Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{tenant_id}/{one_id}/groups":{"get":{"tags":["用户"],"summary":"Get User Groups","operationId":"get_user_groups_users__tenant_id___one_id__groups_get","parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"integer","title":"Tenant Id"}},{"name":"one_id","in":"path","required":true,"schema":{"type":"integer","title":"One Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GroupCreate":{"properties":{"tenant_id":{"type":"integer","title":"Tenant Id"},"group_code":{"type":"string","pattern":"^[a-zA-Z0-9_\\-]+$","title":"Group Code","examples":["vip_high_value"]},"group_name":{"type":"string","title":"Group Name","examples":["VIP高价值用户"]},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"group_type":{"type":"string","title":"Group Type","default":"static","examples":["static","dynamic"]},"filter_rule":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Filter Rule"}},"type":"object","required":["tenant_id","group_code","group_name"],"title":"GroupCreate"},"GroupResponse":{"properties":{"tenant_id":{"type":"integer","title":"Tenant Id"},"group_id":{"type":"integer","title":"Group Id"},"group_code":{"type":"string","title":"Group Code"},"group_name":{"type":"string","title":"Group Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"group_type":{"type":"string","title":"Group Type"},"filter_rule":{"title":"Filter Rule"},"member_count":{"type":"integer","title":"Member Count"},"created_at":{"title":"Created At"},"updated_at":{"title":"Updated At"}},"type":"object","required":["tenant_id","group_id","group_code","group_name","group_type","member_count"],"title":"GroupResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"scale_tier":{"type":"string","title":"Scale Tier"},"doris_be_shards":{"type":"integer","title":"Doris Be Shards"},"executor":{"type":"object","title":"Executor"},"nl_query":{"type":"object","title":"Nl Query"}},"type":"object","required":["status","scale_tier","doris_be_shards","executor","nl_query"],"title":"HealthResponse"},"MemberAdd":{"properties":{"one_id":{"type":"integer","title":"One Id","examples":[100001]},"source":{"type":"string","title":"Source","default":"manual"}},"type":"object","required":["one_id"],"title":"MemberAdd"},"MemberResponse":{"properties":{"tenant_id":{"type":"integer","title":"Tenant Id"},"group_id":{"type":"integer","title":"Group Id"},"one_id":{"type":"integer","title":"One Id"},"added_at":{"title":"Added At"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},"type":"object","required":["tenant_id","group_id","one_id"],"title":"MemberResponse"},"NlPlan":{"properties":{"template":{"type":"string","title":"Template"},"params":{"type":"object","title":"Params"},"source":{"type":"string","title":"Source"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"}},"type":"object","required":["template","params","source"],"title":"NlPlan"},"NlQueryRequest":{"properties":{"question":{"type":"string","title":"Question","examples":["查手机号13800138001的用户画像"]},"tenant_id":{"type":"integer","title":"Tenant Id","examples":[1001]},"context":{"type":"object","title":"Context"}},"type":"object","required":["question","tenant_id"],"title":"NlQueryRequest"},"NlQueryResponse":{"properties":{"question":{"type":"string","title":"Question"},"tenant_id":{"type":"integer","title":"Tenant Id"},"plan":{"$ref":"#/components/schemas/NlPlan"},"result":{"$ref":"#/components/schemas/QueryResult"}},"type":"object","required":["question","tenant_id","plan","result"],"title":"NlQueryResponse"},"QueryRequest":{"properties":{"params":{"type":"object","title":"Params","examples":[{"one_id":100001,"tenant_id":1001}]}},"type":"object","title":"QueryRequest"},"QueryResult":{"properties":{"template":{"type":"string","title":"Template"},"params":{"type":"object","title":"Params"},"row_count":{"type":"integer","title":"Row Count"},"elapsed_ms":{"type":"number","title":"Elapsed Ms"},"data":{"items":{"type":"object"},"type":"array","title":"Data"}},"type":"object","required":["template","params","row_count","elapsed_ms","data"],"title":"QueryResult"},"UserSummary":{"properties":{"one_id":{"type":"integer","title":"One Id"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"wechat_openid":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Wechat Openid"},"tags":{"title":"Tags"},"properties":{"title":"Properties"},"groups":{"items":{"type":"string"},"type":"array","title":"Groups"}},"type":"object","required":["one_id"],"title":"UserSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"系统","description":"健康检查"},{"name":"模板查询","description":"预置 SQL 模板 + 参数拼装"},{"name":"自然语言","description":"DeepSeek 语义理解 → 模板查询"},{"name":"用户分组","description":"人群包 CRUD 与成员管理"},{"name":"用户","description":"用户画像与分组归属查询"}]}